Add shellscript to install my favorite programs
When setting up a new system it can be annoying to spend 30 - 50
minutes installing programs. This script automates that process
as much as possible by adding PPAs I trust and installing programs
that I like to use. This specific script is for Ubuntu 18.04
systems since that distro is the one I use the most.